v3

+++ c/coeff_var.c
@@ -40,7 +40,7 @@ int coeff_var_s (float *Y, float *X, const size_t R, const size_t C,const size_t
         const int ldb = (iscolmajor) ? (int)N2 : 1;
         const int ldc = (iscolmajor) ? (int)R : (int)C;
         float *x1;
-        if (!(x1=(float *)malloc((size_t)N1*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(x1=(float *)malloc(N1*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
         cblas_scopy((int)N1,&ni,0,x1,1);
         cblas_sgemv(Ord,Tr,(int)R,(int)C,1.0f,X,ldc,x1,1,0.0f,Y,1);
         cblas_scopy((int)N1,&o,0,x1,1);
@@ -53,7 +53,7 @@ int coeff_var_s (float *Y, float *X, const size_t R, const size_t C,const size_t
         else
         {
             float *mn;
-            if (!(mn=(float *)malloc((size_t)N2*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
+            if (!(mn=(float *)malloc(N2*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
             cblas_scopy((int)N2,Y,1,mn,1);
             for (size_t n=0; n<N; n++) { X[n] *= X[n]; }
             cblas_sgemv(Ord,Tr,(int)R,(int)C,den,X,ldc,x1,1,0.0f,Y,1);
@@ -64,8 +64,8 @@ int coeff_var_s (float *Y, float *X, const size_t R, const size_t C,const size_t
     else if (iscolmajor && dim==0)
     {
         float *x1, *xni;
-        if (!(x1=(float *)malloc((size_t)R*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
-        if (!(xni=(float *)malloc((size_t)R*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(x1=(float *)malloc(R*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(xni=(float *)malloc(R*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
         cblas_scopy((int)R,&o,0,x1,1); cblas_scopy((int)R,&ni,0,xni,1);
         for (size_t h=0; h<H; h++)
         {
@@ -81,8 +81,8 @@ int coeff_var_s (float *Y, float *X, const size_t R, const size_t C,const size_t
     else if (iscolmajor && dim==1)
     {
         float *x1, *xni;
-        if (!(x1=(float *)malloc((size_t)C*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
-        if (!(xni=(float *)malloc((size_t)C*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(x1=(float *)malloc(C*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(xni=(float *)malloc(C*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
         cblas_scopy((int)C,&o,0,x1,1); cblas_scopy((int)C,&ni,0,xni,1);
         for (size_t h=0; h<H; h++)
         {
@@ -98,8 +98,8 @@ int coeff_var_s (float *Y, float *X, const size_t R, const size_t C,const size_t
     else if (!iscolmajor && dim==2 && H==1)
     {
         float *x1, *xni;
-        if (!(x1=(float *)malloc((size_t)N1*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
-        if (!(xni=(float *)malloc((size_t)N1*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(x1=(float *)malloc(N1*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(xni=(float *)malloc(N1*sizeof(float)))) { fprintf(stderr,"error in coeff_var_s: problem with malloc. "); perror("malloc"); return 1; }
         cblas_scopy((int)N1,&o,0,x1,1); cblas_scopy((int)N1,&ni,0,xni,1);
         for (size_t r=0; r<R; r++)
         {
@@ -154,7 +154,7 @@ int coeff_var_d (double *Y, double *X, const size_t R, const size_t C,const size
         const int ldb = (iscolmajor) ? (int)N2 : 1;
         const int ldc = (iscolmajor) ? (int)R : (int)C;
         double *x1;
-        if (!(x1=(double *)malloc((size_t)N1*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(x1=(double *)malloc(N1*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
         cblas_dcopy((int)N1,&ni,0,x1,1);
         cblas_dgemv(Ord,Tr,(int)R,(int)C,1.0,X,ldc,x1,1,0.0,Y,1);
         cblas_dcopy((int)N1,&o,0,x1,1);
@@ -167,7 +167,7 @@ int coeff_var_d (double *Y, double *X, const size_t R, const size_t C,const size
         else
         {
             double *mn;
-            if (!(mn=(double *)malloc((size_t)N2*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
+            if (!(mn=(double *)malloc(N2*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
             cblas_dcopy((int)N2,Y,1,mn,1);
             for (size_t n=0; n<N; n++) { X[n] *= X[n]; }
             cblas_dgemv(Ord,Tr,(int)R,(int)C,den,X,ldc,x1,1,0.0,Y,1);
@@ -178,8 +178,8 @@ int coeff_var_d (double *Y, double *X, const size_t R, const size_t C,const size
     else if (iscolmajor && dim==0)
     {
         double *x1, *xni;
-        if (!(x1=(double *)malloc((size_t)R*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
-        if (!(xni=(double *)malloc((size_t)R*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(x1=(double *)malloc(R*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(xni=(double *)malloc(R*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
         cblas_dcopy((int)R,&o,0,x1,1); cblas_dcopy((int)R,&ni,0,xni,1);
         for (size_t h=0; h<H; h++)
         {
@@ -195,8 +195,8 @@ int coeff_var_d (double *Y, double *X, const size_t R, const size_t C,const size
     else if (iscolmajor && dim==1)
     {
         double *x1, *xni;
-        if (!(x1=(double *)malloc((size_t)C*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
-        if (!(xni=(double *)malloc((size_t)C*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(x1=(double *)malloc(C*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(xni=(double *)malloc(C*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
         cblas_dcopy((int)C,&o,0,x1,1); cblas_dcopy((int)C,&ni,0,xni,1);
         for (size_t h=0; h<H; h++)
         {
@@ -212,8 +212,8 @@ int coeff_var_d (double *Y, double *X, const size_t R, const size_t C,const size
     else if (!iscolmajor && dim==2 && H==1)
     {
         double *x1, *xni;
-        if (!(x1=(double *)malloc((size_t)N1*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
-        if (!(xni=(double *)malloc((size_t)N1*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(x1=(double *)malloc(N1*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(xni=(double *)malloc(N1*sizeof(double)))) { fprintf(stderr,"error in coeff_var_d: problem with malloc. "); perror("malloc"); return 1; }
         cblas_dcopy((int)N1,&o,0,x1,1); cblas_dcopy((int)N1,&ni,0,xni,1);
         for (size_t r=0; r<R; r++)
         {
