+++ c/ctranspose.c
@@ -111,7 +111,7 @@ int ctranspose_inplace_s (float *X, const size_t R, const size_t C, const char i
     else
     {
         float *Xt;
-        if (!(Xt=(float *)malloc((size_t)(R*C)*sizeof(float)))) { fprintf(stderr,"error in ctranspose_inplace_s: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(Xt=(float *)malloc(R*C*sizeof(float)))) { fprintf(stderr,"error in ctranspose_inplace_s: problem with malloc. "); perror("malloc"); return 1; }
         cblas_scopy((int)(R*C),X,1,Xt,1);
         if (iscolmajor)
         {
@@ -143,7 +143,7 @@ int ctranspose_inplace_d (double *X, const size_t R, const size_t C, const char 
     else
     {
         double *Xt;
-        if (!(Xt=(double *)malloc((size_t)(R*C)*sizeof(double)))) { fprintf(stderr,"error in ctranspose_inplace_d: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(Xt=(double *)malloc(R*C*sizeof(double)))) { fprintf(stderr,"error in ctranspose_inplace_d: problem with malloc. "); perror("malloc"); return 1; }
         cblas_dcopy((int)(R*C),X,1,Xt,1);
         if (iscolmajor)
         {
@@ -173,7 +173,7 @@ int ctranspose_inplace_c (float *X, const size_t R, const size_t C, const char i
     else
     {
         float *Xt;
-        if (!(Xt=(float *)malloc((size_t)(2*R*C)*sizeof(float)))) { fprintf(stderr,"error in ctranspose_inplace_c: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(Xt=(float *)malloc(2*R*C*sizeof(float)))) { fprintf(stderr,"error in ctranspose_inplace_c: problem with malloc. "); perror("malloc"); return 1; }
         cblas_ccopy((int)(R*C),X,1,Xt,1);
         cblas_sscal((int)(R*C),-1.0f,&Xt[1],2);
         if (iscolmajor)
@@ -204,7 +204,7 @@ int ctranspose_inplace_z (double *X, const size_t R, const size_t C, const char 
     else
     {
         double *Xt;
-        if (!(Xt=(double *)malloc((size_t)(2*R*C)*sizeof(double)))) { fprintf(stderr,"error in ctranspose_inplace_z: problem with malloc. "); perror("malloc"); return 1; }
+        if (!(Xt=(double *)malloc(2*R*C*sizeof(double)))) { fprintf(stderr,"error in ctranspose_inplace_z: problem with malloc. "); perror("malloc"); return 1; }
         cblas_zcopy((int)(R*C),X,1,Xt,1);
         cblas_dscal((int)(R*C),-1.0,&Xt[1],2);
         if (iscolmajor)
